refactor: move store to the src directory, change name of a selector in bookSlice
