+++ src/context/index.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from 'app/store';
+import store from 'store';
 import { QueryClientProvider } from './query-client';
 
 function AppProviders({ children }) {
